+++ ninja-servlet/src/main/java/ninja/servlet/ContextImpl.java
@@ -24,6 +24,7 @@ import java.net.URI;
 import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -46,6 +47,10 @@ import ninja.servlet.async.AsyncStrategy;
 import ninja.servlet.async.AsyncStrategyFactoryHolder;
 import ninja.session.FlashScope;
 import ninja.session.Session;
+import ninja.uploads.FileItem;
+import ninja.uploads.FileItemProvider;
+import ninja.uploads.FileProvider;
+import ninja.uploads.NoFileItemProvider;
 import ninja.utils.HttpHeaderUtils;
 import ninja.utils.NinjaConstant;
 import ninja.utils.NinjaProperties;
@@ -55,13 +60,18 @@ import ninja.utils.SwissKnife;
 import ninja.validation.Validation;
 
 import org.apache.commons.fileupload.FileItemIterator;
+import org.apache.commons.fileupload.FileItemStream;
 import org.apache.commons.fileupload.FileUploadException;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
+import org.apache.commons.fileupload.util.Streams;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.common.collect.ArrayListMultimap;
+import com.google.common.collect.Multimap;
 import com.google.inject.Inject;
+import com.google.inject.Injector;
 
 public class ContextImpl implements Context.Impl {
 
@@ -85,6 +95,11 @@ public class ContextImpl implements Context.Impl {
     private final Session session;
     private final ResultHandler resultHandler;
     private final Validation validation;
+    private final Injector injector;
+    
+    private boolean formFieldsProcessed = false;
+    private Map<String, List<String>> formFieldsMap;
+    private Map<String, List<FileItem>> fileFieldsMap;
 
     // In Async mode, these values will be set to null, so save them
     private String requestPath;
@@ -93,13 +108,14 @@ public class ContextImpl implements Context.Impl {
     private Logger logger = LoggerFactory.getLogger(ContextImpl.class);
 
     @Inject
-    public ContextImpl(
+    public ContextImpl (
             BodyParserEngineManager bodyParserEngineManager,
             FlashScope flashCookie,
             NinjaProperties ninjaProperties,
             ResultHandler resultHandler,
             Session sessionCookie,
-            Validation validation) {
+            Validation validation,
+            Injector injector) {
 
         this.bodyParserEngineManager = bodyParserEngineManager;
         this.flashScope = flashCookie;
@@ -107,6 +123,7 @@ public class ContextImpl implements Context.Impl {
         this.session = sessionCookie;
         this.resultHandler = resultHandler;
         this.validation = validation;
+        this.injector = injector;
     }
 
     public void init(ServletContext servletContext,
@@ -123,7 +140,7 @@ public class ContextImpl implements Context.Impl {
 
         // init session scope:
         session.init(this);
-
+        
         contextPath = httpServletRequest.getContextPath();
         requestPath = performGetRequestPath();
     }
@@ -164,16 +181,29 @@ public class ContextImpl implements Context.Impl {
 
     @Override
     public String getParameter(String key) {
-        return httpServletRequest.getParameter(key);
+        if (!formFieldsProcessed) processFormFields();
+        if (formFieldsMap == null) {
+            return httpServletRequest.getParameter(key);
+        } else {
+            List<String> values = formFieldsMap.get(key);
+            if (values == null || values.isEmpty())
+                return null;
+            return values.get(0);
+        }
     }
 
     @Override
     public List<String> getParameterValues(String name) {
-        String[] params = httpServletRequest.getParameterValues(name);
-        if (params == null) {
-            return Collections.emptyList();
+        if (!formFieldsProcessed) processFormFields();
+        if (formFieldsMap == null) {
+            String[] params = httpServletRequest.getParameterValues(name);
+            if (params == null) {
+                return Collections.emptyList();
+            }
+            return Arrays.asList(params);
+        } else {
+            return formFieldsMap.get(name);
         }
-        return Arrays.asList(params);
     }
 
     @Override
@@ -202,7 +232,32 @@ public class ContextImpl implements Context.Impl {
             return defaultValue;
         }
     }
+    
+    @Override
+    public FileItem getParameterAsFileItem(String key) {
+        if (!formFieldsProcessed) processFormFields();
+        if (fileFieldsMap == null) return null;
+        List<FileItem> fileItems = fileFieldsMap.get(key);
+        if (fileItems == null || fileItems.isEmpty()) return null;
+        return fileItems.get(0);
+    }
+    
+    @Override
+    public List<FileItem> getParameterAsFileItems(String key) {
+        if (!formFieldsProcessed) processFormFields();
+        if (fileFieldsMap == null) return Collections.emptyList();
+        List<FileItem> fileItems = fileFieldsMap.get(key);
+        if (fileItems == null) return Collections.emptyList();
+        return fileItems;
+    }
 
+    @Override
+    public Map<String, List<FileItem>> getParameterFileItems() {
+        if (!formFieldsProcessed) processFormFields();
+        return fileFieldsMap;
+    }
+
+    
     @Override
     public <T> T getParameterAs(String key, Class<T> clazz) {
         return getParameterAs(key, clazz, null);
@@ -221,7 +276,18 @@ public class ContextImpl implements Context.Impl {
 
     @Override
     public Map<String, String[]> getParameters() {
-        return httpServletRequest.getParameterMap();
+        if (!formFieldsProcessed) processFormFields();
+        if (formFieldsMap == null) {
+            return httpServletRequest.getParameterMap();
+        } else {
+            // convert List<String> value to String[] value
+            String[] type = new String[0];
+            Map<String, String[]> map = new HashMap<>(formFieldsMap.size());
+            for (Entry<String, List<String>> entry: formFieldsMap.entrySet()) {
+                map.put(entry.getKey(), entry.getValue().toArray(type));
+            }
+            return map;
+        }
     }
 
     @Override
@@ -322,6 +388,11 @@ public class ContextImpl implements Context.Impl {
     @Override
     public List<Cookie> getCookies() {
         javax.servlet.http.Cookie[] servletCookies = httpServletRequest.getCookies();
+        
+        if (servletCookies == null) {
+            return Collections.EMPTY_LIST;
+        }
+        
         List<Cookie> ninjaCookies = new ArrayList<>(servletCookies.length);
 
         for (javax.servlet.http.Cookie cookie : servletCookies) {
@@ -578,7 +649,13 @@ public class ContextImpl implements Context.Impl {
     @Override
     public FileItemIterator getFileItemIterator() {
 
+        long maxFileSize = ninjaProperties.getIntegerWithDefault(NinjaConstant.UPLOADS_MAX_FILE_SIZE, -1);
+        long maxTotalSize = ninjaProperties.getIntegerWithDefault(NinjaConstant.UPLOADS_MAX_TOTAL_SIZE, -1);
+        
         ServletFileUpload upload = new ServletFileUpload();
+        upload.setFileSizeMax(maxFileSize);
+        upload.setSizeMax(maxTotalSize);
+        
         FileItemIterator fileItemIterator = null;
 
         try {
@@ -633,6 +710,22 @@ public class ContextImpl implements Context.Impl {
     @Override
     public void setAttribute(String name, Object value) {
         httpServletRequest.setAttribute(name, value);
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        // build map of attributes
+        Map<String,Object> attributes = new HashMap<String,Object>();
+        
+        Enumeration<String> en = httpServletRequest.getAttributeNames();
+        
+        while (en.hasMoreElements()) {
+            String name = en.nextElement();
+            Object value = httpServletRequest.getAttribute(name);
+            attributes.put(name, value);
+        }
+        
+        return attributes;
     }
 
     /**
@@ -714,5 +807,109 @@ public class ContextImpl implements Context.Impl {
         }
 
         return contentType.startsWith(ContentTypes.APPLICATION_XML);
+    }
+    
+    private void processFormFields() {
+        if (formFieldsProcessed) return;
+        formFieldsProcessed = true;
+
+        // return if not multipart
+        if (!ServletFileUpload.isMultipartContent(httpServletRequest))
+            return;
+
+        // get fileProvider from route method/class, or defaults to an injected one
+        // if none injected, then we do not process form fields this way and let the user
+        // call classic getFileItemIterator() by himself
+        FileProvider fileProvider = null;
+        if (route != null) {
+            if (fileProvider == null) {
+                fileProvider = route.getControllerMethod().getAnnotation(FileProvider.class);
+            }
+            if (fileProvider == null) {
+                fileProvider = route.getControllerClass().getAnnotation(FileProvider.class);
+            }
+        }
+        
+        // get file item provider from file provider or default one
+        FileItemProvider fileItemProvider = null;
+        if (fileProvider == null) {
+            fileItemProvider = injector.getInstance(FileItemProvider.class);
+        } else {
+            fileItemProvider = injector.getInstance(fileProvider.value());
+        }
+        
+        if (fileItemProvider instanceof NoFileItemProvider) return;
+        
+        // Initialize maps and other constants
+        ArrayListMultimap<String, String> formMap = ArrayListMultimap.create();
+        ArrayListMultimap<String, FileItem> fileMap = ArrayListMultimap.create();
+        
+        
+        // This is the iterator we can use to iterate over the contents of the request.
+        try {
+            
+            FileItemIterator fileItemIterator = getFileItemIterator();
+            
+            while (fileItemIterator.hasNext()) {
+
+                FileItemStream item = fileItemIterator.next();
+
+                if (item.isFormField()) {
+                    
+                    // save the form field for later use from getParameter
+                    String value = Streams.asString(item.openStream());
+                    formMap.put(item.getFieldName(), value);
+
+                } else {
+                    
+                    // process file as input stream and save for later use in getParameterAsFile or getParameterAsInputStream
+                    FileItem fileItem = fileItemProvider.create(item);
+                    fileMap.put(item.getFieldName(), fileItem);
+                }
+            }
+        } catch (FileUploadException | IOException e) {
+            throw new RuntimeException("Failed to parse multipart request data", e);
+        }
+
+        // convert both multimap<K,V> to map<K,List<V>>
+        formFieldsMap = toUnmodifiableMap(formMap);
+        fileFieldsMap = toUnmodifiableMap(fileMap);
+    }
+    
+    /**
+     * Utility method to convert a Guava Multimap to an unmodifiable Map that
+     * uses a List<T> as a value. Optimized for the case where values are already
+     * internally stored as a List<T> (e.g. ArrayListMultimap).
+     * @param <T> The value type
+     * @param multimap The multimap to convert from
+     * @return The unmodifiable converted map
+     */
+    private <T> Map<String, List<T>> toUnmodifiableMap(Multimap<String, T> multimap) {
+        Map<String, List<T>> map = new HashMap<>(multimap.size());
+        
+        for (Entry<String, Collection<T>> entry: multimap.asMap().entrySet()) {
+            Collection<T> value = entry.getValue();
+            if (value == null) {
+                Collections.emptyList();
+            } else if (value instanceof List) {
+                map.put(entry.getKey(), (List<T>)value);
+            } else {
+                map.put(entry.getKey(), new ArrayList<>(value));
+            }
+        }
+        
+        return Collections.unmodifiableMap(map);
+    }
+    
+    @Override
+    public void cleanup() {
+        // call cleanup on all file items
+        if (fileFieldsMap != null) {
+            for (List<FileItem> files: fileFieldsMap.values()) {
+                for (FileItem file: files) {
+                    file.cleanup();
+                }
+            }
+        }
     }
 }