+++ ninja-test-utilities/src/main/java/ninja/utils/FakeContext.java
@@ -32,6 +32,7 @@ import ninja.Result;
 import ninja.Route;
 import ninja.session.FlashScope;
 import ninja.session.Session;
+import ninja.uploads.FileItem;
 import ninja.validation.Validation;
 import ninja.validation.ValidationImpl;
 
@@ -446,6 +447,11 @@ public class FakeContext implements Context {
         attributes.put(name, value);
     }
 
+    @Override
+    public Map<String, Object> getAttributes() {
+        return attributes;
+    }
+    
 	@Override
 	public String getContextPath() {
 		return contextPath;
@@ -483,5 +489,24 @@ public class FakeContext implements Context {
     @Override
     public void unsetCookie(Cookie cookie) {
         throw new UnsupportedOperationException("Not supported in fake context");        
+    }
+
+    @Override
+    public FileItem getParameterAsFileItem(String name) {
+        throw new UnsupportedOperationException("Not supported in fake context");        
+    }
+
+    @Override
+    public List<FileItem> getParameterAsFileItems(String name) {
+        throw new UnsupportedOperationException("Not supported in fake context");        
+    }
+
+    @Override
+    public Map<String, List<FileItem>> getParameterFileItems() {
+        throw new UnsupportedOperationException("Not supported in fake context");        
+    }
+
+    @Override
+    public void cleanup() {
     }
 }