+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,6 +1,36 @@
-Version X.X.X
+Versoin X.X.X
 =============
 
+ * 2015-10-18 OptionalBinder for ObjectMapper and XmlMapper, so that users can easily override
+              the default types for both of them. (amit2103/jjlauer)
+
+Version 5.2.1
+=============
+
+ * 2015-10-16 Context.getParameterFileItems() now returns Map<String, List<FileItem>> (jjlauer)
+
+Version 5.2.0
+=============
+
+ * 2015-08-17 Added gzip-enabled rollover appender examples to archetypes' logback.xml 
+              files & minor doc typo fixes (metacity)
+ * 2015-06-24 Injection of params and uploaded files from multipart requests (bazi, momiji)
+
+Version 5.1.7
+=============
+
+ * 2015-10-06 AuthenticityFilter uses Ninja interface not NinjaDefault class (jjlauer)
+ * 2015-10-06 AuthenticityFilter logs filtered requests as warnings (jjlauer)
+ * 2015-10-01 Fixed garbled snippets in diagnostics (mallowlabs)
+ * 2015-09-25 Upgrade FreeMarker from 2.3.22 to 2.3.23 (khmarbaise)
+ * 2015-09-28 Improved diagnostic mode w/ info about the context, request, and response (jjlauer)
+ * 2015-09-28 Fixed NPE issue with cookies not being set in underlying servlet request (jjlauer)
+ * 2015-09-28 NinjaDefault now logs exceptions in the debug level (jjlauer)
+
+Version 5.1.6
+=============
+
+ * 2015-09-17 Fixed bug in AssetsController that occurred in dev mode on some environments (ra).
  * 2015-08-23 Added charset to error html files (mallowlabs)
  * 2015-09-15 Fixed status code for forbidden results in diagnostic mode (jjlauer)
  * 2015-09-15 Added documentation for session configuration properties (jjlauer)